// src/features/master-job/types/service.types.ts
export interface ServiceProduct {
    id: string
    product_code: string
    product_name: string
    alias: string | null
    descriptions: string
    product_price: string
    status_product: string
    created_by: string
    created_at: string
    update_at: string
    update_by: string
    photo: string
    status_price: string
    type_price: string
    kena_pajak: string
    als_coa_id: string
}